django_backend/api_app/serializers.py
# serializers.py
from rest_framework import serializers
from .models import SensorData

class SensorDataSerializer(serializers.ModelSerializer):
    class Meta:
        model = SensorData
        fields = ['sensor_id', 'temperature', 'humidity', 'timestamp']